Change to include license info in generated gemspec.

+++ lib/tty/commands/new.rb
@@ -10,7 +10,6 @@ require 'open3'
 module TTY
   module Commands
     class New < Cmd
-
       # @api private
       attr_reader :app_name
 
@@ -20,6 +19,8 @@ module TTY
       # @api private
       attr_reader :target_path
 
+      attr_reader :templates
+
       def initialize(app_path, options)
         @app_path = app_path
         @app_name = resolve_name(app_path)
@@ -27,6 +28,7 @@ module TTY
         @pastel   = Pastel.new
 
         @target_path = Pathname.pwd.join(@app_path)
+        @templates   = {}
       end
 
       def resolve_name(name)
@@ -34,7 +36,8 @@ module TTY
       end
 
       def template_source_path
-        ::File.expand_path(::File.join(::File.dirname(__FILE__), '..', 'templates/new'))
+        path = ::File.join(::File.dirname(__FILE__), '..', 'templates/new')
+        ::File.expand_path(path)
       end
 
       def git_exist?
@@ -48,16 +51,32 @@ module TTY
       def template_options
         opts = OpenStruct.new
         opts[:app_name] = app_name,
-        opts[:author]   = git_author.empty? ? "TODO: Write your name" : git_author
+        opts[:author]   = git_author.empty? ? 'TODO: Write your name' : git_author
         opts
       end
 
-      def templates
-        @templates ||= { }
+      def licenses
+        @licenses ||= {
+          'agplv3' => { name: 'AGPL-3.0',
+                        desc: 'GNU Affero General Public License v3.0' },
+          'apache' => { name: 'Apache-2.0', desc: 'Apache License 2.0' },
+          'gplv2'  => { name: 'GPL-2.0',
+                        desc: 'GNU General Public License v2.0' },
+          'gplv3'  => { name: 'GPL-3.0',
+                        desc: 'GNU General Public License v3.0' },
+          'lgplv3' => { name: 'LGPL-3.0',
+                        desc: 'GNU Lesser General Public License v3.0' },
+          'mit'    => { name: 'MIT', desc: 'MIT License' },
+          'mplv2'  => { name: 'MPL-2.0', desc: 'Mozilla Public License 2.0' }
+        }
       end
 
       def add_mapping(source, target)
-        templates.merge!(source => target)
+        @templates[source] = target
+      end
+
+      def gemspec_path
+        target_path.join("#{app_name}.gemspec").to_s
       end
 
       # Execute the command
@@ -69,18 +88,26 @@ module TTY
 
         coc_opt  = options['coc'] ? '--coc' : '--no-coc'
         test_opt = options['test']
-        command = "bundle gem #{target_path} --no-mit --no-exe #{coc_opt} -t #{test_opt}"
-        out, _ = run(command)
+        command = [
+          "bundle gem #{target_path}",
+          '--no-mit',
+          '--no-exe',
+          coc_opt,
+          "-t #{test_opt}"
+        ].join(' ')
+
+        out, = run(command)
 
         if !options['no-color']
-          out = out.gsub(/^(\s+)(create)/, '\1' + @pastel.green('\2')).
-                    gsub(/^(\s+)(identical)/, '\1' + @pastel.yellow('\2'))
+          out = out.gsub(/^(\s+)(create)/, '\1' + @pastel.green('\2'))
+                   .gsub(/^(\s+)(identical)/, '\1' + @pastel.yellow('\2'))
         end
 
         license = options['license'] == 'none' ? false : options['license']
 
         if license
-          add_mapping("#{license}_LICENSE.txt.erb", "LICENSE.txt")
+          add_mapping("#{license}_LICENSE.txt.erb", 'LICENSE.txt')
+          add_license_to_gemspec(license)
         end
 
         puts out
@@ -88,8 +115,30 @@ module TTY
         templates.each do |src, dst|
           source = ::File.join(template_source_path, src)
           destination = target_path.join(dst).to_s
+          next unless ::File.exist?(source)
           copy_file(source, destination, context: template_options)
         end
+      end
+
+      def add_license_to_gemspec(license)
+        gemspec = ::File.read(gemspec_path)
+        gemspec_var_name = gemspec.match(/(\w+)\.name/)[1]
+        matches = gemspec.match(/^(\s*)#{gemspec_var_name}\.name(\s*)=.*$/)
+        gemspec_pre_indent  = matches[1].size
+        gemspec_post_indent = matches[2].size - ('license'.size - 'name'.size)
+
+        license_regex = /(^\s*#{gemspec_var_name}\.license\s*=\s*).*/
+        if gemspec =~ license_regex
+          gemspec.gsub!(liecense_regex, "\\1\"#{licenses[license][:name]}\"")
+        else
+          gem_license = ' ' * gemspec_pre_indent
+          gem_license << "#{gemspec_var_name}.license"
+          gem_license << ' ' * gemspec_post_indent
+          gem_license << "= \"#{licenses[license][:name]}\""
+          gemspec.gsub!(/(^\s*#{gemspec_var_name}\.name\s*=\s*.*$)/,
+                        "\\1\n#{gem_license}")
+        end
+        ::File.write(gemspec_path, gemspec)
       end
     end # New
   end # Commands
